requirements.txt
attrs==19.3.0
botocore==1.17.0
certifi==2020.4.5.2
chardet==3.0.4
docutils==0.15.2
idna==2.9
importlib-metadata==1.6.1
jmespath==0.10.0
marshmallow==3.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
pynamodb==4.3.2
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-decouple==3.3
requests==2.23.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
